requirements.txt
Flask~=1.1.2
Flask-JWT-Extended~=3.13.1
Flask-RESTful
Flask-SQLAlchemy
bcrypt~=3.2.0
pytz~=2021.1
datetime~=4.3
tzlocal~=2.1
psycopg2
PyJWT~=1.7.1
requests~=2.25.1
Werkzeug~=1.0.1
SQLAlchemy~=1.3.23
python-dotenv~=0.15.0